+++ __tests__/ManageKey.test.ts
@@ -1,11 +1,12 @@
 import {
   AuthorizeAlias,
+  CommunicationsEnum,
   FetchKey,
   GeneratePacket,
   ServerResponse,
 } from "../src";
-
 import { ensureCredentialsPresent, sdk, testEmail } from "./utils/setupFiles";
+import { IGeneratePacketParams } from "../src/com/xqmsg/sdk/v2/services/GeneratePacket";
 
 describe("Testing manual key management", () => {
   const authorizeUser = (user: string) =>
@@ -19,16 +20,18 @@ describe("Testing manual key management", () => {
             return response;
           }
           case ServerResponse.ERROR: {
-            return false;
+            return response;
           }
         }
       });
 
-  const payload = {
+  const payload: IGeneratePacketParams = {
     [GeneratePacket.KEY]: "TEST KEY",
     [GeneratePacket.EXPIRES_HOURS]: 1,
     [GeneratePacket.RECIPIENTS]: ["recipient@xqmsg.com"],
     [GeneratePacket.DELETE_ON_RECEIPT]: false,
+    [GeneratePacket.TYPE]: CommunicationsEnum.EMAIL,
+    [GeneratePacket.META]: null,
   };
 
   const generatePacket = () =>
@@ -50,9 +53,6 @@ describe("Testing manual key management", () => {
     new FetchKey(sdk)
       .supplyAsync({ [FetchKey.LOCATOR_KEY]: token })
       .then((response) => {
-        if (!response) {
-          return false;
-        }
         switch (response.status) {
           case ServerResponse.OK: {
             return response.payload;
@@ -68,9 +68,6 @@ describe("Testing manual key management", () => {
   it(`should successfully generate and fetch key packet`, async () =>
     expect(
       await authorizeUser(testEmail).then((userResponse) => {
-        if (!userResponse) {
-          return false;
-        }
         switch (userResponse.status) {
           case ServerResponse.OK: {
             return generatePacket().then((keyResponse) => {