+++ src/com/xqmsg/sdk/v2/services/GetSubscriberInfo.ts
@@ -2,13 +2,16 @@ import CallMethod from "../CallMethod";
 import ServerResponse from "../ServerResponse";
 import XQModule from "./XQModule";
 import XQSDK from "../XQSDK";
+import { XQServices } from "../XQServicesEnum";
+
+import handleException from "../exceptions/handleException";
 
 /**
- * A service which is utilized to retrieve a users information.
+ * A service which is utilized to retrieve the current subscriber's information.
  *
- * @class [GetUserInfo]
+ * @class [GetSubscriberInfo]
  */
-export default class GetUserInfo extends XQModule {
+export default class GetSubscriberInfo extends XQModule {
   /** The required fields of the payload needed to utilize the service */
   requiredFields: string[];
 
@@ -56,24 +59,29 @@ export default class GetUserInfo extends XQModule {
           Authorization: "Bearer " + accessToken,
         };
 
-        return this.sdk.call(
-          this.sdk.SUBSCRIPTION_SERVER_URL,
-          this.serviceName,
-          CallMethod.GET,
-          additionalHeaderProperties,
-          maybePayLoad,
-          true
-        );
+        return this.sdk
+          .call(
+            this.sdk.SUBSCRIPTION_SERVER_URL,
+            this.serviceName,
+            CallMethod.GET,
+            additionalHeaderProperties,
+            maybePayLoad,
+            true
+          )
+          .then((response: ServerResponse) => {
+            switch (response.status) {
+              case ServerResponse.OK: {
+                return response;
+              }
+              case ServerResponse.ERROR: {
+                return handleException(response, XQServices.GetSubscriberInfo);
+              }
+            }
+          });
       } catch (exception) {
-        return new Promise((resolve) => {
-          resolve(
-            new ServerResponse(
-              ServerResponse.ERROR,
-              exception.code,
-              exception.reason
-            )
-          );
-        });
+        return new Promise((resolve) =>
+          resolve(handleException(exception, XQServices.GetSubscriberInfo))
+        );
       }
     };
   }