+++ src/com/xqmsg/sdk/v2/services/RevokeUserAccess.ts
@@ -2,6 +2,9 @@ import CallMethod from "../CallMethod";
 import ServerResponse from "../ServerResponse";
 import XQModule from "./XQModule";
 import XQSDK from "../XQSDK";
+import { XQServices } from "../XQServicesEnum";
+
+import handleException from "../exceptions/handleException";
 
 /**
  * A service which is utilized to revoke access to keys for specific recipients without revoking the entire token.
@@ -62,24 +65,29 @@ export default class RevokeUserAccess extends XQModule {
           Authorization: "Bearer " + accessToken,
         };
 
-        return this.sdk.call(
-          this.sdk.VALIDATION_SERVER_URL,
-          this.serviceName + "/" + encodeURIComponent(locatorKey),
-          CallMethod.OPTIONS,
-          additionalHeaderProperties,
-          payload,
-          true
+        return this.sdk
+          .call(
+            this.sdk.VALIDATION_SERVER_URL,
+            this.serviceName + "/" + encodeURIComponent(locatorKey),
+            CallMethod.PATCH,
+            additionalHeaderProperties,
+            payload,
+            true
+          )
+          .then((response: ServerResponse) => {
+            switch (response.status) {
+              case ServerResponse.OK: {
+                return response;
+              }
+              case ServerResponse.ERROR: {
+                return handleException(response, XQServices.RevokeUserAccess);
+              }
+            }
+          });
+      } catch (exception) {
+        return new Promise((resolve) =>
+          resolve(handleException(exception, XQServices.RevokeUserAccess))
         );
-      } catch (validationException) {
-        return new Promise((resolve) => {
-          resolve(
-            new ServerResponse(
-              ServerResponse.ERROR,
-              validationException.code,
-              validationException.reason
-            )
-          );
-        });
       }
     };
   }