+++ src/com/xqmsg/sdk/v2/services/dashboard/RemoveUserGroup.ts
@@ -3,6 +3,9 @@ import Destination from "../../Destination";
 import ServerResponse from "../../ServerResponse";
 import XQModule from "../XQModule";
 import XQSDK from "../../XQSDK";
+import { XQServices } from "../../XQServicesEnum";
+
+import handleException from "../../exceptions/handleException";
 
 /**
  * A service which is utilized to remove a grouping of dashboard users
@@ -42,25 +45,30 @@ export default class RemoveUserGroup extends XQModule {
           Authorization: "Bearer " + dashboardAccessToken,
         };
 
-        return this.sdk.call(
-          this.sdk.DASHBOARD_SERVER_URL,
-          this.serviceName + "/" + maybePayLoad[RemoveUserGroup.ID],
-          CallMethod.DELETE,
-          additionalHeaderProperties,
-          null,
-          true,
-          Destination.DASHBOARD
-        );
+        return this.sdk
+          .call(
+            this.sdk.DASHBOARD_SERVER_URL,
+            this.serviceName + "/" + maybePayLoad[RemoveUserGroup.ID],
+            CallMethod.DELETE,
+            additionalHeaderProperties,
+            null,
+            true,
+            Destination.DASHBOARD
+          )
+          .then(async (response: ServerResponse) => {
+            switch (response.status) {
+              case ServerResponse.OK: {
+                return response;
+              }
+              case ServerResponse.ERROR: {
+                return handleException(response, XQServices.RemoveUserGroup);
+              }
+            }
+          });
       } catch (exception) {
-        return new Promise((resolve) => {
-          resolve(
-            new ServerResponse(
-              ServerResponse.ERROR,
-              exception.code,
-              exception.reason
-            )
-          );
-        });
+        return new Promise((resolve) =>
+          resolve(handleException(exception, XQServices.RemoveUserGroup))
+        );
       }
     };
   }