+++ src/com/xqmsg/sdk/v2/services/dashboard/VerifyAccount.ts
@@ -9,54 +9,77 @@ import { XQServices } from "../../XQServicesEnum";
 
 import handleException from "../../exceptions/handleException";
 
+enum DashboardAccessToken {
+  preAuth = "preauth-dashboard",
+  auth = "dashboard",
+}
+
 /**
- * A service utilized to log-in a user and allow access to Dashboard services.
+ * A service utilized to verify a user via their `accessToken` and allow access to Dashboard services.
  * @class [DashboardLogin]
  */
-export default class DashboardLogin extends XQModule {
+export default class VerifyAccount extends XQModule {
   /** The required fields of the payload needed to utilize the service */
   requiredFields: string[];
 
   /** Specified name of the service */
   serviceName: string;
 
-  /** The field name representing the service name */
-  static LOGIN: "login" = "login";
-
-  /** The field name representing the authenticated pwd */
-  static PWD: "pwd" = "pwd";
+  /** The field name representing an access token */
+  static ACCESS_TOKEN: "accessToken" = "accessToken";
 
   /**
    * @param {Map} maybePayLoad - the container for the request parameters supplied to this method.
-   * @param {String} maybePayLoad.pwd - the provided OAuth token
+   * @param {String} maybePayLoad.accesstoken - the provided access token
    * @returns {Promise<ServerResponse<{payload:string}>>}
    */
-  supplyAsync: (maybePayLoad: {
-    [DashboardLogin.PWD]: string;
+  supplyAsync: (maybePayload: {
+    [VerifyAccount.ACCESS_TOKEN]: string;
   }) => Promise<ServerResponse>;
 
   constructor(sdk: XQSDK) {
     super(sdk);
-    this.serviceName = DashboardLogin.LOGIN;
-    this.requiredFields = [DashboardLogin.PWD];
+    this.serviceName = "login/verify";
+    this.requiredFields = [VerifyAccount.ACCESS_TOKEN];
 
     this.supplyAsync = (maybePayLoad) => {
       try {
         const self = this;
-        this.sdk.validateInput(maybePayLoad, this.requiredFields);
 
-        const loginRequest = {
-          method: 1, // TODO(worstestes - 3.21.22): an obselete field which will be removed at a later date
-          [DashboardLogin.PWD]: maybePayLoad.pwd,
+        const accessToken = maybePayLoad[VerifyAccount.ACCESS_TOKEN];
+
+        const additionalHeaderProperties = {
+          Authorization: `Bearer ${accessToken}`,
         };
+
+        const decodedIncomingAccessToken: JwtPayload = jwtDecode(accessToken);
+
+        // if user has an already existing dashboard token
+        // skip pre-auth dashboard token exchange process
+        if (decodedIncomingAccessToken.iss === DashboardAccessToken.auth) {
+          const profile = decodedIncomingAccessToken.sub;
+
+          self.cache.putActiveProfile(profile);
+
+          self.cache.putDashboardAccess(profile, accessToken);
+
+          return new Promise((resolve) => {
+            resolve(
+              new ServerResponse(ServerResponse.OK, 200, {
+                user: profile,
+                dashboardAccessToken: accessToken,
+              })
+            );
+          });
+        }
 
         return this.sdk
           .call(
             this.sdk.DASHBOARD_SERVER_URL,
             this.serviceName,
-            CallMethod.POST,
+            CallMethod.GET,
+            additionalHeaderProperties,
             null,
-            loginRequest,
             true,
             Destination.DASHBOARD
           )
@@ -78,19 +101,20 @@ export default class DashboardLogin extends XQModule {
                   activeProfile,
                   dashboardAccessToken
                 );
+
                 return new ServerResponse(ServerResponse.OK, 200, {
                   user: profile,
                   dashboardAccessToken,
                 });
               }
               case ServerResponse.ERROR: {
-                return handleException(response, XQServices.DashboardLogin);
+                return handleException(response, XQServices.VerifyAccount);
               }
             }
           });
       } catch (exception) {
         return new Promise((resolve) =>
-          resolve(handleException(exception, XQServices.DashboardLogin))
+          resolve(handleException(exception, XQServices.VerifyAccount))
         );
       }
     };